requirements.txt
streamlit>=1.36
biopython>=1.81
numpy>=1.24
pandas>=2.0
scikit-learn==1.3.2
joblib>=1.3
